refactor(annict): 連携を別ページからprofileインラインへ
別ページ(app/annict.tsx)への遷移はUX的に埋もれるため廃止し、
連携状態+連携/解除を AnnictConnectionCard として profile タブに
インライン表示する。カードは将来のソフトゲート(PR6)でも再利用可能。

- app/annict.tsx と _layout.tsx の Stack 登録を削除
- components/AnnictConnectionCard.tsx を新規追加
- 解除確認のキャンセルに annict.cancel キーを追加(ja/en)
