+++ apps/mobile/app/_layout.tsx
@@ -39,7 +39,6 @@ function AuthGuard() {
       <Stack.Screen name="(auth)" />
       <Stack.Screen name="(tabs)" />
       <Stack.Screen name="user" />
-      <Stack.Screen name="annict" />
     </Stack>
   );
 }
