+++ apps/mobile/lib/i18n/locales/ja/translation.json
@@ -11,6 +11,7 @@
     "disconnect": "連携を解除する",
     "connected": "Annict と連携済み",
     "notConnected": "未連携",
+    "cancel": "キャンセル",
     "softGate": {
       "title": "Annict 連携が必要です",
       "description": "視聴記録の表示・更新には Annict との連携が必要です。",
